requirements.txt
Flask
Flask-Caching
pandas
pandas_ta
yfinance>=0.2.40
requests
plotly
textblob
apscheduler
python-dotenv
gunicorn
curl_cffi
feedparser
